ui: add style counter
