+++ lib/views/navigation/index.dart
@@ -104,132 +104,6 @@ class _AddImageDropRegionState extends State<AddImageDropRegion> {
     }
 }
 
-class DefaultDrawer extends StatelessWidget {
-    const DefaultDrawer({super.key, this.activeView, this.displayTitle = true, this.desktopView = false});
-
-    final bool displayTitle;
-    final bool desktopView;
-    final String? activeView;
-
-    @override
-    Widget build(context) {
-        Color? assertSelected(String selectorView) {
-            if(activeView == selectorView) return Theme.of(context).colorScheme.primary;
-        }
-        return ListView(
-            padding: EdgeInsets.zero,
-            children: <Widget>[
-                SizedBox(height: MediaQuery.of(context).viewPadding.top),
-                if(displayTitle) const Padding(
-                    padding: EdgeInsets.all(16),
-                    child: Text("LocalBooru",
-                        style: TextStyle(
-                            fontSize: 20.0
-                        )
-                    ),
-                ),
-                if(desktopView) ...[
-                    ListTile(
-                        textColor: assertSelected("home"),
-                        iconColor: assertSelected("home"),
-                        title: const Text("Home"),
-                        leading: const Icon(Icons.home),
-                        onTap: activeView != "home" ? () {
-                            Scaffold.of(context).closeDrawer();
-                            context.go("/home");
-                        } : null,
-                    ),
-                    ListTile(
-                        textColor: assertSelected("recent") ?? assertSelected("search"),
-                        iconColor: assertSelected("recent") ?? assertSelected("search"),
-                        title: const Text("Search"),
-                        leading: const Icon(Icons.search),
-                        onTap: activeView != "recent" && activeView != "search" ? () {
-                            Scaffold.of(context).closeDrawer();
-                            context.push("/recent");
-                        } : null,
-                    ),
-                    const Divider(),
-                ],
-                ListTile(
-                    textColor: assertSelected("manage_image"),
-                    iconColor: assertSelected("manage_image"),
-                    title: const Text("Add image"),
-                    leading: const Icon(Icons.add),
-                    onTap: activeView != "manage_image" ? () {
-                        Scaffold.of(context).closeDrawer();
-                        context.push("/manage_image");
-                    } : null,
-                ),
-                ListTile(
-                    title: const Text("Import from service"),
-                    leading: const Icon(Icons.link),
-                    enabled: activeView != "manage_image",
-                    onTap: () {
-                        Scaffold.of(context).closeDrawer();
-                        showDialog(
-                            context: context,
-                            builder: (context) {
-                                return const InsertURLDialog();
-                            },
-                        );
-                    },
-                ),
-                ListTile(
-                    textColor: assertSelected("settings"),
-                    iconColor: assertSelected("settings"),
-                    title: const Text("Settings"),
-                    leading: const Icon(Icons.settings),
-                    onTap: activeView != "settings" ? () {
-                        Scaffold.of(context).closeDrawer();
-                        context.push("/settings");
-                    } : null,
-                ),
-                if(kDebugMode) ...[
-                    const Divider(),
-                    const Padding(
-                        padding: EdgeInsets.only(left: 16.0, top: 16.0),
-                        child: Text("Dev mode"),
-                    ),
-                    ListTile(
-                        title: const Text("Playground"),
-                        onTap: () {
-                            Scaffold.of(context).closeDrawer();
-                            context.push("/playground");
-                        },
-                    ),
-                    ListTile(
-                        title: const Text("Go to permissions screen"),
-                        onTap: () {
-                            Scaffold.of(context).closeDrawer();
-                            context.push("/permissions");
-                        },
-                    ),
-                    ListTile(
-                        title: const Text("Go to set booru screen"),
-                        onTap: () {
-                            Scaffold.of(context).closeDrawer();
-                            context.push("/setbooru");
-                        },
-                    ),
-                    ListTile(
-                        title: const Text("Desktop size"),
-                        onTap: () {
-                            appWindow.size = const Size(1280, 720);
-                        },
-                    ),
-                    ListTile(
-                        title: const Text("Phone size"),
-                        onTap: () {
-                            appWindow.size = const Size(320, 840);
-                        },
-                    ),
-                ]
-            ],
-        );
-    }
-}
-
 class BrowseScreenPopupMenuButton extends StatelessWidget {
     const BrowseScreenPopupMenuButton({super.key, this.image});
